Initial commit from skeleton
I ran the following command to generate the skeleton

mdbook init git-ps-book

and at the prompt I gave it a title of "Git Patch Stack Book" and said
yes when asked if I wanted it to include a .gitignore.

This will hopefully get a basic thing up and running so I can start
porting content over to it and getting it put together.