Add hooks entry for list_additional_info
Add documentation for the new feature to add additional info to the list command through a hook

This is done so users of gps know about the hook and how to use it.

I added an entry to the hooks page about the existing hooks.

Relates to https://github.com/uptech/git-ps-rs/issues/123

Should only be integrated after a version is released with the feature
[changelog]
added: hook entry for list_additional_info

ps-id: ff1bcae9-bab2-4b46-9066-94a61944b9e4